Workaround for posts collection conflict
The EveryPolitician posts collection conflicts with Jekyll's posts
collection. In the longer term we'll probably need to namespace
collections or something, but as a short term fix I'm just treating the
posts collection as if it's a Jekyll collection and using the top level
`permalink` option to specify the correct url structure.

If/when this site gets a blog then this workaround will become
problematic, but hopefully by that time it will be fixed at a different
level.