Fix missing/broken images
Previously we were incorrectly assuming that if the person had an
'image' property then there would be a mirror image on the GitHub Pages
CDN. This isn't true as not all images have been successfully retrieved
from the parliament site.

This changes things so we explicitly check the index.txt file for a
person's id before setting their image, and we explicitly delete the
image property if they don't have one available.